Fix tests on Pyton 3.5.
In Python 3.6 `json.loads()` was changed to accept both strings and bytes.
Previously, it accepted only strings.  Werkzeug's Request and Response
method get_data() by default returns bytes (as does the `data`
descriptor that calls it, which we previously used).  This breaks tests
when running on Python 3.5, so we need to `get_data(as_text=True)`.